+++ src/datasources/UserDataSource.ts
@@ -3,7 +3,7 @@ import { ProposerPayload } from '../types/proposal';
 import { UserUpdationEventPayload } from '../types/user';
 
 export interface UserDataSource {
-  create(user: ProposerPayload): Promise<User>;
+  create(user: ProposerPayload): Promise<User | null>;
   update(user: UserUpdationEventPayload): Promise<User>;
   delete(userId: number): Promise<number>;
 }
