+++ src/datasources/postgres/ExperimentUserDataSource.ts
@@ -1,25 +1,15 @@
-import { Knex } from 'knex';
-import { container } from 'tsyringe';
-
-import { Tokens } from '../../config/Tokens';
 import { ExperimentUser } from '../../models/ExperimentUser';
 import {
   ExperimentUserRecord,
   createExperimentUserObject,
 } from '../../types/records';
 import { ExperimentUserDataSource } from '../ExperimentUserDataSource';
-import Database from './database/index';
+import database from './database';
 export default class PostgresExperimentUserDataSource
   implements ExperimentUserDataSource
 {
   private TABLE_NAME = 'experiment_user';
-  private database: Knex;
-  constructor() {
-    const databaseInstance = container.resolve<Database>(Tokens.Database);
-    (async () => {
-      this.database = await databaseInstance.connect();
-    })();
-  }
+
   async create({
     experimentId,
     userId,
@@ -27,7 +17,7 @@ export default class PostgresExperimentUserDataSource
     experimentId: string;
     userId: string;
   }): Promise<ExperimentUser> {
-    const experimentUserExists = await this.database(this.TABLE_NAME).where({
+    const experimentUserExists = await database(this.TABLE_NAME).where({
       experiment_id: experimentId,
       user_id: userId,
     });
@@ -36,7 +26,7 @@ export default class PostgresExperimentUserDataSource
       return experimentUserExists[0];
     }
 
-    return await this.database(this.TABLE_NAME)
+    return await database(this.TABLE_NAME)
       .insert({
         experiment_id: experimentId,
         user_id: userId,
