+++ src/datasources/postgres/UserDataSource.ts
@@ -1,7 +1,3 @@
-import { Knex } from 'knex';
-import { container } from 'tsyringe';
-
-import { Tokens } from '../../config/Tokens';
 import { Employer } from '../../models/Employer';
 import { User } from '../../models/User';
 import { ProposerPayload } from '../../types/proposal';
@@ -13,23 +9,16 @@ import {
 } from '../../types/records';
 import { UserUpdationEventPayload } from '../../types/user';
 import { UserDataSource } from '../UserDataSource';
-import Database from './database/index';
+import database from './database';
 export default class PostgresUserDataSource implements UserDataSource {
   private TABLE_NAME = 'users';
   private EMPLOYER_TABLE_NAME = 'employer';
   private USER_ROLE = 'user_role';
   private ROLE = 'role';
 
-  private database: Knex;
-  constructor() {
-    const databaseInstance = container.resolve<Database>(Tokens.Database);
-    (async () => {
-      this.database = await databaseInstance.connect();
-    })();
-  }
-
-  async create(user: ProposerPayload): Promise<User> {
-    const userExists = await this.database(this.TABLE_NAME)
+  async create(user: ProposerPayload): Promise<User | null> {
+    if (!user.email || !user.institution) return null;
+    const userExists = await database(this.TABLE_NAME)
       .where({
         id: user.email,
       })
@@ -39,7 +28,7 @@ export default class PostgresUserDataSource implements UserDataSource {
       });
     // Create an Employer, if it does not exist
     let employer: Employer;
-    const employerExists = await this.database(this.EMPLOYER_TABLE_NAME)
+    const employerExists = await database(this.EMPLOYER_TABLE_NAME)
       .where({
         id: user.institution.id,
       })
@@ -47,11 +36,10 @@ export default class PostgresUserDataSource implements UserDataSource {
       .then((employer: EmployerRecord) => {
         return employer ? createEmployerObject(employer) : null;
       });
-
     if (employerExists) {
       employer = employerExists;
     } else {
-      employer = await this.database(this.EMPLOYER_TABLE_NAME)
+      employer = await database(this.EMPLOYER_TABLE_NAME)
         .insert({
           id: user.institution.id,
           name: user.institution.name,
@@ -65,7 +53,7 @@ export default class PostgresUserDataSource implements UserDataSource {
     if (userExists) {
       // Update user table if the institution_id is different
       if (userExists.affiliationId !== employer.id) {
-        return await this.database(this.TABLE_NAME)
+        return await database(this.TABLE_NAME)
           .where({
             email: user.email,
           })
@@ -81,7 +69,7 @@ export default class PostgresUserDataSource implements UserDataSource {
       }
     } else {
       // Insert into users table
-      return await this.database(this.TABLE_NAME)
+      return await database(this.TABLE_NAME)
         .insert({
           id: user.email,
           email: user.email,
@@ -92,14 +80,14 @@ export default class PostgresUserDataSource implements UserDataSource {
         })
         .returning(['*'])
         .then(async (user: UserRecord[]) => {
-          const scientificComputingRole = await this.database(this.ROLE)
+          const scientificComputingRole = await database(this.ROLE)
             .where({
               name: 'SCIENTIFIC_COMPUTING',
             })
             .first();
 
           if (scientificComputingRole) {
-            await this.database(this.USER_ROLE).insert({
+            await database(this.USER_ROLE).insert({
               user_id: user[0].id,
               role_id: scientificComputingRole.id,
             });
@@ -111,13 +99,13 @@ export default class PostgresUserDataSource implements UserDataSource {
   }
 
   async update(user: UserUpdationEventPayload): Promise<User> {
-    const userExists = await this.database(this.TABLE_NAME).where({
+    const userExists = await database(this.TABLE_NAME).where({
       email: user.email,
     });
 
     // Update only if the User exists and submitted
     if (userExists) {
-      return await this.database(this.TABLE_NAME)
+      return await database(this.TABLE_NAME)
         .where({
           email: user.email,
         })
@@ -133,7 +121,7 @@ export default class PostgresUserDataSource implements UserDataSource {
   }
 
   async delete(id: number) {
-    await this.database(this.TABLE_NAME)
+    await database(this.TABLE_NAME)
       .where({
         id: id,
       })
