+++ src/datasources/postgres/database/remote.ts
@@ -1,12 +1,12 @@
-import { Knex } from 'knex';
+import knex from 'knex';
 
-import { connectWithVisa } from '../../../synchronisation';
+const conString = `postgres://${process.env.VISA_DB_USERNAME}:${
+  process.env.VISA_DB_PASSWORD
+}@127.0.0.1:${9091}/${process.env.VISA_DB_NAME}`;
 
-import Database from './index';
-export default class LocalDatabase implements Database {
-  async connect(): Promise<Knex> {
-    const knex = await connectWithVisa();
+const db = knex({
+  client: 'pg',
+  connection: conString,
+});
 
-    return knex;
-  }
-}
+export default db;
