+++ src/index.ts
@@ -1,7 +1,6 @@
 import 'reflect-metadata';
 import { logger } from '@user-office-software/duo-logger';
 import 'dotenv/config';
-import { logger } from '@user-office-software/duo-logger';
 import express from 'express';
 import { container } from 'tsyringe';
 import './config';
@@ -10,7 +9,9 @@ import { Tokens } from './config/Tokens';
 import healthCheck from './middlewares/healthCheck';
 import readinessCheck from './middlewares/readinessCheck';
 import startQueueHandling from './queue/queueHandling';
-import { synchronization } from './synchronisation';
+import connectWithUserOffice from './synchronisation/ConnectWithUserOffice';
+import connectWithVisa from './synchronisation/ConnectWithVisa';
+import synchronization from './synchronisation/Synchronize';
 
 async function bootstrap() {
   const PORT = process.env.PORT || 4010;
@@ -30,12 +31,19 @@ async function bootstrap() {
   );
 
   container.resolve<() => void>(Tokens.ConfigureLogger)();
-  startQueueHandling();
 
-  app.get('/synchronize', (req, res) => {
-    synchronization();
-    res.send('Hello World!');
-  });
+  if (process.env.DEPENDENCY_CONFIG === 'local') {
+    startQueueHandling();
+  }
+
+  if (process.env.DEPENDENCY_CONFIG === 'local') {
+    await connectWithVisa();
+    await connectWithUserOffice();
+    app.get('/synchronize', async (req, res) => {
+      await synchronization();
+      res.send('---- SynchronisationDone---');
+    });
+  }
 }
 
 bootstrap();
